modificacion atributos venta

+++ src/main/java/org/example/ServicioTecnico.java
@@ -143,7 +143,7 @@ public class ServicioTecnico {
 		return null;
 	}
 
-	public void registrarVenta(String nombreCliente, String apellidoCliente, String fecha, int iva, double total, firebase firebaseInstance) {
+	public void registrarVenta(String nombreCliente, String apellidoCliente, String fecha, double iva, double total, firebase firebaseInstance) {
 		for (Cliente cliente : clientesRut) {
 			String nombre = cliente.getNombre();
 			String apellido = cliente.getApellido();
